replaced onsstove.py
